Remove extension and simplify task configuration
My understanding of an extension in Gradle is that it holds
configuration that applies across a whole project. For example, it may
be used to provide project-wide configuration for several different
tasks of the same type. As such, I don't think it makes sense for this
plugin's extension to offer SourceSet configuration. If the SourceSets
are removed, only the encoding is left. I'm not sure that this one
property, which will almost always be UTF-8, warrants an extension.

This commit removes the extension and also simplifies the plugin's
task configuration.