Make task names more idiomatic
As Gradle tasks get more specific, words are normally added to the
task's name but the ordering of the words remains the same. For
example, there's a clean task that cleans everything, but there are
also more specific cleaning tasks that only clean particular things:

- clean
- cleanEclipse
- cleanEclipseClasspath

This plugin's format checking is run when check is run, but can also
be run individually to perform more focussed checking. Previously, the
names of the format checking tasks all began with "formatcheck" but
this doesn't match the naming convention described above.

This commit updates the format checking tasks so that their names
all being with checkFormat. This means that the relevant task names
now follow the same pattern as the clean tasks listed above:

- check
- checkFormatMain
- checkFormatTest

See gh-22