Add support for "required this" customization
Add `IdentCheckFilter`, `RequiresOuterThisFilter` and `SpringNoThisCheck`
classes which can be used to customize when `this.` is required.

The `IdentCheckFilter` can be used to wrap the standard
`RequiresThisCheck` and skip specifically named idents. The
`RequiresOuterThisFilter` can be used to skip `RequiresThisCheck` on
inner classes and `SpringNoThisCheck` can be used to enforce that
certain idents never use `this.`.

Closes gh-45