Align with Buildship 3.0's API
Previously, the plugin was incorrectly using Buildship 3.1 API while
claiming compatibility with 3.0. This could lead to NoSuchMethodErrors
at runtime.

This commit updates the code to use Buildship 3.0 API. It also
updates the manifest to constrain the version of the
org.eclipse.buildship.core bundle to 3.0.x versions. Lastly, the
Eclipse repository that is used has been rolled back to 2019-03
as 2019-06 bundles Buildship 3.1.

Closes gh-121