Use a ProjectConfigurator extension to configure Gradle projects
Previously, the plugin listened to a Gradle project's settings being
refreshed to apply its settings to the project. This only worked if
the plugin had been previously activated by one of its other
extensions (m2e or code formatting). In situations where the plugin
had not already been activated, the refresh would be missed and the
plugin's settings were not applied to the project.

This commit switches from the listener-based approach to using a
ProjectConfigurator extension. Buildship automatically calls the
extension when a Gradle project is being configured, either because
it's being imported into the workspace for the first time or because
its settings are being refreshed. The use of an extension means that
the plugin is automatically activated if needed when Buildship
calls the extension.

Closes gh-123