Fix shaded formatter prefs
Update the shaded jar so that the key values in `formatter.prefs` are
also processed.

Prior to this commit the keys all started with `org.eclipse` and were
unfortunately not picked up due to the fact that the shade plugin
changed the value of the `JavaCore.PLUGIN_ID` constant in the bytecode.

Closes gh-161