Upgrade to Eclipse 2021-09
2021-09 requires Java 11 so the version of ASM has been upgraded to
one that supports Java 11 bytecode and EclipseRewriter has been
updated to use the matching ASM API version.

2021-09 also contains an updated version of the NLS class that we
rewrite to remove a call to System.getProperty. Previously, this
call was done in an isolated fashion in the run method of an
anonmyous inner-class. It's now been updated [1] to use a lamdba
which results in the call to System.getProperty now being part of
the class's static initializer. This commit rewrites this <clinit>
to retain the existing initialization of EMPTY_ARGS and ASSIGNED
fields while updating the initialization of ignoreWarnings to
replace the System.getProperty call with a hardcoded true value.

Closes gh-277

[1] https://git.eclipse.org/c/equinox/rt.equinox.framework.git/commit/bundles/org.eclipse.osgi/supplement/src/org/eclipse/osgi/util/NLS.java?id=ba6205dab1f857fe87cf5985f03c1b5e88797064