Check that annotation attribute values are concise
When the value of an annotation's attribute references something
defined by an inner type of the annotation, the inner type should be
imported and referenced directly rather than through the containing
type.

Consider the following:

@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)

The repetition of SpringBootTest in both the annotation itself and
the value of one of its attributes is overly verbose. Instead, we
prefer to import WebEnvironment and refer to it directly:

@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)

This commit adds a check that enforces the style described above. In
the event of there being an existing import for a different
WebEnvironment type, no error will be reported as the more verbose
style is necessary to avoid referencing the wrong WebEnvironment.

Closes gh-452